Support http proxies during authorization
Enabled SystemConfiguration in QNetworkProxyFactory in management.py as
well as initializing oauth2.Client's proxy settings.

Moved the proxy helper from everpad.provider.tools to everpad.tools,
since the oath stuff is in the indicator, not in the pad, and getting the
proxy info from the environment should happen in both.

Fixes #156

Test Plan: Configured my system to read through a localhost HTTP CONNECT
proxy (https://github.com/fmoo/twisted-connect-proxy), verified that the
authorize flow was sending both the initial OAuth handshake, as well as
the HTTP requests from the QWebView, etc,  through the proxy.
